@@ -1,16 +1,15 @@
 apply plugin: 'com.android.library'
-apply plugin: 'android-apt'
-apply plugin: 'bintray-release'
+apply plugin: 'com.novoda.bintray-release'
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.2"
+    compileSdkVersion 'android-28'
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 21
+        targetSdkVersion 28
         versionCode 1
-        versionName "1.0"
+        versionName '1.0'
     }
 
     lintOptions {
@@ -28,8 +27,6 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.jakewharton.fliptables:fliptables:1.0.2'
 }
 
 publish {
