@@ -1,8 +1,8 @@
 package com.example.android.kvs;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import com.example.android.kvs.prefs.schemas.TestWithDefaultPrefs;
 
